\name{family}
\docType{data}
\alias{family}
\title{Data set family}
\description{
Data set to be used as example for demo purpose.
}
\usage{
family
}
\author{
\packageAuthor{wyz.code.rdoc}

Maintainer: \packageMaintainer{wyz.code.rdoc}
}
\keyword{datasets}
\concept{documentation generation}
\encoding{UTF-8}
\source{
Data set generated by NEONIRA
}
\section{Classification}{
STRATUM \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} LAYER_3

PHASING \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} BUILD

INTENT  \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} UTILITIES
}
