\name{generateParagraphCR}
\alias{generateParagraphCR}
\title{Function generateParagraphCR}
\description{
Data set to be used as example for demo purpose.
}
\usage{
generateParagraphCR(..., addFinalSeparator_b_1 = FALSE)
}
\arguments{
\item{...}{additional arguments.}
\item{addFinalSeparator_b_1}{A single boolean value}
}
\author{
\packageAuthor{wyz.code.rdoc}

Maintainer: \packageMaintainer{wyz.code.rdoc}
}
\seealso{Functions \code{\link{generateParagraph}} and
\code{\link{generateParagraph2NL}}.}
\keyword{function}
\concept{documentation generation}
\encoding{UTF-8}
\source{
Data set generated by NEONIRA
}
\section{Classification}{
STRATUM \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} LAYER_1

PHASING \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} BUILD

INTENT  \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} PARTS_BUILDING
}
\examples{
generateParagraphCR(LETTERS[1:3])
# "A\\\\crB\\\\crC"

generateParagraphCR(LETTERS[1:3], addFinalSeparator_b_1 = TRUE)
# "A\\\\crB\\\\crC\\cr"
}
