\name{rdocKeywords}
\alias{rdocKeywords}
\title{R Documentation Keywords}
\description{Provides all \R documentation markup tags a.k.a keywords}
\usage{
rdocKeywords(asList_b_1 = FALSE)
}
\arguments{
\item{asList_b_1}{A single boolean value. }
}
\value{A vector of type characters, containing all keywords when parameter
\code{asList_b_1} is \code{FALSE}. Otherwise a \code{list} organizing this same
content. }
\references{
Refer to
\href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing \R extensions}.
}
\author{
\packageAuthor{wyz.code.rdoc}

Maintainer: \packageMaintainer{wyz.code.rdoc}
}
\keyword{documentation}
\concept{documentation generation}
\encoding{UTF-8}
\section{Classification}{
STRATUM \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} CORE

PHASING \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} RUN

INTENT  \ifelse{latex}{\out{$\longmapsto$}}{\ifelse{html}{\out{&#x25B6;}}{\\u25b6}} FEEDBACK
}
\examples{
rdocKeywords()
}
