\name{WrapperObject}
\alias{WrapperObject}
\title{Wrapper object.}
\description{
Class \code{WrapperObject} from package \code{wyz.code.metaTesting}.
}
\usage{
WrapperObject()
}
\value{
An object of class \code{WrapperObject}.
}
\note{
This object is intended for programmers, not for end-users.
}
\author{
\packageAuthor{wyz.code.rdoc}

Maintainer: \packageMaintainer{wyz.code.rdoc}
}
\examples{
# typical use case
wo <- WrapperObject()
}
