\name{generateParagraphCR}
\alias{generateParagraphCR}
\title{Function generateParagraphCR}
\description{
Data set to be used as example for demo purpose.
}
\usage{
generateParagraphCR(..., addFinalSeparator_b_1 = FALSE)
}
\arguments{
\item{...}{additional arguments.}
\item{addFinalSeparator_b_1}{A single \emph{\code{boolean}} value}
}
\value{A single \emph{\code{string}}, with possibly many embedded '\\cr' character sequences.}
\author{
\packageAuthor{wyz.code.rdoc}

Maintainer: \packageMaintainer{wyz.code.rdoc}
}
\seealso{Functions \emph{\bold{\code{\link{generateParagraph}}}} and
\emph{\bold{\code{\link{generateParagraph2NL}}}}.}
\keyword{function}
\concept{documentation generation}
\encoding{UTF-8}
\source{
Data set generated by NEONIRA
}
\examples{
generateParagraphCR(LETTERS[1:3])
# "A\\\\crB\\\\crC"

generateParagraphCR(LETTERS[1:3], addFinalSeparator_b_1 = TRUE)
# "A\\\\crB\\\\crC\\cr"
}
