\name{opRdocInformation}
\alias{opRdocInformation}
\title{Offensive Programming R Documentation Information}
\description{List package functions and provide informations about their intented usage. }
\usage{
opRdocInformation()
}
\value{
See \bold{\emph{\code{\link[wyz.code.offensiveProgramming]{opInformation}}}} value
description.
}
\author{
\packageAuthor{wyz.code.rdoc}

Maintainer: \packageMaintainer{wyz.code.rdoc}
}
\keyword{documentation}
\concept{documentation generation}
\encoding{UTF-8}
\examples{
opRdocInformation()
}